requirements.txt
streamlit
langchain
openai
python-dotenv
faiss-cpu
langchain-community
tiktoken
pypdf